SYstem MVP FUlly fullfilled

+++ openGLPhysics/opengl-physics/tools/include/base_camera.h
@@ -18,8 +18,8 @@ namespace tools
 
 		glm::mat4 get_projection() const;
 
-		virtual bool event_key(Direction dir, double deltaTime) = 0;
-		virtual bool event_key(double deltaTime, float xMove, float yMove) = 0;
+		virtual bool event_key(Direction dir, float deltaTime) = 0;
+		virtual bool event_key(float deltaTime, float xMove, float yMove) = 0;
 		virtual bool event_scroll(float yOffset) = 0;
 
 	protected:
